requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
openai-whisper==20231117
torch==2.1.0
numpy==1.24.3
lifxlan==1.2.8
websockets==12.0